+++ @utahdts/utah-design-system-header/src/js/settings/defaultSettings.js
@@ -24,6 +24,7 @@ export default {
   },
   showTitle: true,
   size: sizes.MEDIUM,
+  skipLinkUrl: '#main-content',
   title: 'My Utah.gov Site',
   titleURL: '/',
   utahId: false,