+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [N.E.xT] 12/13/2023
+## Added
+- Header: Main Menu - self menu items
+
+## Fixed
+- Table overflow horizontal scroll bar
+
 # [1.12.0] 12/13/2023
 ## Added
 - Table Pagination component that syncs table filtering and sorting